+++ src/server/app.ts
@@ -5,10 +5,10 @@ import * as helmet from "helmet";
 import { join } from "path";
 import * as serveStatic from "serve-static";
 import * as uuid from "uuid/v4";
-import { useLimiter } from "../limiter";
 import { prepareApi } from "./api";
 import { prepareAuth } from "./auth";
 import { Env } from "./env";
+import { useLimiter } from "./limiter";
 import { authRouter } from "./routes/auth";
 import { fallbackRouter } from "./routes/fallback";
 import { logoutRouter } from "./routes/logout";
